+++ mne/io/eeglab/eeglab.py
@@ -242,7 +242,7 @@ class RawEEGLAB(_BaseRaw):
     def _read_segment_file(self, data, idx, fi, start, stop, cals, mult):
         """Read a chunk of raw data"""
         _read_segments_file(self, data, idx, fi, start, stop, cals, mult,
-                            dtype=np.float32, n_bytes=4)
+                            dtype=np.float32)
 
 
 class EpochsEEGLAB(_BaseEpochs):