+++ mne/io/utils.py
@@ -127,17 +127,17 @@ def _blk_read_lims(start, stop, buf_len):
 
 
 def _read_segments_file(raw, data, idx, fi, start, stop, cals, mult,
-                        dtype='<i2', n_bytes=2):
+                        dtype='<i2'):
     """Read a chunk of raw data"""
     n_channels = raw.info['nchan']
-
+    n_bytes = np.dtype(dtype).itemsize
     # data_offset and data_left count data samples (channels x time points),
     # not bytes.
     data_offset = n_channels * start * n_bytes
     data_left = (stop - start) * n_channels
 
     # Read up to 100 MB of data at a time, block_size is in data samples
-    block_size = ((100e6 // n_bytes) // n_channels) * n_channels
+    block_size = ((int(100e6) // n_bytes) // n_channels) * n_channels
     block_size = min(data_left, block_size)
     with open(raw._filenames[fi], 'rb', buffering=0) as fid:
         fid.seek(data_offset)
@@ -147,7 +147,7 @@ def _read_segments_file(raw, data, idx, fi, start, stop, cals, mult,
             count = min(block_size, data_left - sample_start * n_channels)
             block = np.fromfile(fid, dtype, count)
             block = block.reshape(n_channels, -1, order='F')
-            n_samples = block.shape[1]
+            n_samples = block.shape[1]  # = count // n_channels
             sample_stop = sample_start + n_samples
             data_view = data[:, sample_start:sample_stop]
             _mult_cal_one(data_view, block, idx, cals, mult)